Add basic SVG Elements
This commit adds a React wrapper for the SVG element <circle> and
<path>.  The Path wrapper only exposes a subset of the functionality
of <path>.  In particular one must pass a raw data array
`d = [[x0,y0], [x1,y1],..]` to the Path component and the component
will do the job of converting it into a string compatible with svg path
data.  Currently there is no option to overide this behavior and the
Path component will only do open linear paths.  This will likely be
extended in the future.